A simple continuous rapid maths (calculation) based game, which is easy to play.
This is a python based game which runs in command line interface. Batch file is used to trigger the python file and make it run in cmd on windows. Also batch file provides a shortcut to be created of which you can change the icon to the ico provided, designed by me or you can design your own better icon and set it with shortcut for better representation.